Factor out platform schema
