+++ src/uwtools/resources/fv3.jsonschema
@@ -285,37 +285,6 @@
       ],
       "type": "object"
     },
-    "platform": {
-      "additionalProperties": false,
-      "dependentRequired": {
-        "account": [
-          "scheduler"
-        ],
-        "scheduler": [
-          "account"
-        ]
-      },
-      "properties": {
-        "account": {
-          "type": "string"
-        },
-        "mpicmd": {
-          "type": "string"
-        },
-        "scheduler": {
-          "enum": [
-            "lsf",
-            "pbs",
-            "slurm"
-          ],
-          "type": "string"
-        }
-      },
-      "required": [
-        "mpicmd"
-      ],
-      "type": "object"
-    },
     "preprocessing": {
       "additionalProperties": false,
       "properties": {
