+++ src/uwtools/resources/platform.jsonschema
@@ -1,0 +1,35 @@
+{
+  "properties": {
+    "platform": {
+      "additionalProperties": false,
+      "dependentRequired": {
+        "account": [
+          "scheduler"
+        ],
+        "scheduler": [
+          "account"
+        ]
+      },
+      "properties": {
+        "account": {
+          "type": "string"
+        },
+        "mpicmd": {
+          "type": "string"
+        },
+        "scheduler": {
+          "enum": [
+            "lsf",
+            "pbs",
+            "slurm"
+          ],
+          "type": "string"
+        }
+      },
+      "required": [
+        "mpicmd"
+      ],
+      "type": "object"
+    }
+  }
+}
